fixed login logic, resolved errors and warnings
